Remove Clojure dep, add profiles for testing
Removes Clojure itself as a primary dependency, since consumers will
specify the version of Clojure required by their application. Separate
profiles have been added that specify a different version of
Clojure (starting with version 1.5.1 which was being used before this
change), so that you can run `lein with-profile clj-1.5 test` to run
tests with version 1.5.1, for example. This commit adds a simple println
to the current core_test.clj file to make it clear that a different
Clojure version is being exercised with the different profiles.