Fix a parsing bug for project-id
Sentry.io has started to give out project numbers that are bigger than
an integer. This patch changes the parsing of the project id to a
long, instead of an integer.

In addition, the few dependencies that exists are upgraded to the
latest version, with a special mention to clj-http-lite. The
clj-commons version is used there, as it is maintained.