Strip out boot and replace with leiningen
Due to difficulties with getting boot to run, a switch to leiningen
has been made. Leiningen was picked over the official CLI tools from
Clojure as it's super stable and works well with the existing code
base.

A small helper function was added in project.clj to grab the latest version.