Add `clojure.lang.ExceptionInfo`'s `ex-data` where available
`prone.stacks/normalize-exception`, via `prone.stacks/add-data` sets
a `clojure.lang.ExceptionInfo`'s `ex-data` as `:data` on the normalized
exception.  This data is generally useful, so pass it on to Sentry via
the `mechanism.data` key of the `exception` payload:
* https://develop.sentry.dev/sdk/event-payloads/exception/#exception-mechanism

Since Sentry does not support nested maps (it will not display the data at all
in that case), we flatten the data using `pr-str`.